codes/requirements.txt
adce==1.3.3.2
beautifulsoup4==4.11.1
certifi==2020.6.20
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cssselect==1.1.0
cvxopt==1.3.0
cvxpy==1.2.0
cycler==0.11.0
dateparser==1.1.1
ecos==2.0.10
et-xmlfile==1.1.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.7.0
fonttools==4.33.3
GoogleNews==1.6.2
gpkit==1.1.1
idna==3.3
jieba3k==0.35.1
joblib==1.1.0
kiwisolver==1.4.2
lxml==4.8.0
matplotlib==3.5.2
newspaper3k==0.2.8
nltk==3.7
numpy==1.22.3
openpyxl==3.0.10
osqp==0.6.2.post5
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
Pint==0.19.2
plotly==5.7.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qdldl==0.1.5.post2
regex==2022.3.2
requests==2.27.1
requests-file==1.5.1
scipy==1.8.0
scs==3.2.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
tenacity==8.0.1
tinysegmenter==0.3
tldextract==3.3.0
tqdm==4.64.0
tzdata==2022.1
tzlocal==2.1
urllib3==1.26.9
wincertstore==0.2
